// src/app/store/types/transaction.types.ts
export enum TransactionActionTypes {
  LOAD_TRANSACTIONS = '[TRANSACTION] Load Transaction',
  LOAD_TRANSACTIONS_SUCCESS = '[TRANSACTION] Load Transaction Success',
  LOAD_TRANSACTIONS_FAILURE = '[TRANSACTION] Load Transaction Failure',

  ADD_TRANSACTION = '[TRANSACTION] Add Transaction',
  ADD_TRANSACTION_SUCCESS = '[TRANSACTION] Add Transaction Success',
  ADD_TRANSACTION_FAILURE = '[TRANSACTION] Add Transaction Failure',

  DELETE_TRANSACTION = '[TRANSACTION] Delete Transaction',
  DELETE_TRANSACTION_SUCCESS = '[TRANSACTION] Delete Transaction Success',
  DELETE_TRANSACTION_FAILURE = '[TRANSACTION] Delete Transaction Failure',
}